Improve --make and --lesson options
Type: fix

Previously when run like 'new --make=c-project MyMakefile', it created
two files: 'Makefile' that was the c-project template and 'MyMakefile'
that was just empty. I fixed that and now if you call it like that, it
will create only one file 'MyMakefile' - the template. If you omit the
name of file, 'new' will create a default template file 'Makefile'.
The same principle works with '--lesson' option.